chore(openai): add openai base class
